requirements.txt
streamlit==1.41.1
python-dotenv 
google-generativeai==0.8.4
pdf2image 
pillow 
matplotlib 
wordcloud
